requirements.txt
bibtexparser==1.4.0
boto3==1.26.135
dj-database-url==2.0.0
dj-rest-auth==4.0.0
Django==4.2.2
django-configurations==2.4.1
django-compressor==4.4
django-cors-headers==4.2.0
django-currentuser @ https://github.com/zsoldosp/django-currentuser/archive/refs/heads/master.zip
django-elasticsearch-dsl==7.3
django-filter==21.1
django-maintenance-mode==0.18.0
django-recaptcha==3.0.0
django-storages==1.13.2
django_structlog==6.0.0
django-tenants==3.5.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.2.2
djangosaml2idp==0.7.2
drf-access-policy==1.5.0
drf-writable-nested==0.7.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
geopy==2.3.0
gunicorn==20.1.0
lxml==4.9.2
mysqlclient==2.1.1
pandas==2.0.1
passlib==1.7.4
psycopg2==2.9.6
pycurl==7.45.2
pyzotero==1.5.9
requests==2.30.0
requests-aws4auth==1.2.3
stringcase==1.2.0
urllib3==1.26.15
wagtail==5.0
wagtail-modelchooser==4.0.1
Wand===0.6.11
Werkzeug==2.3.4
